WARNING: risk of epilepsy
Very very simple code that just change the color of background like crazy.
Meanwhile, I'm thinking of a way to ellegantly set data in the NES
header—a nesconfig macro that sets the mapper, memory, mirroring etc.

Ah, and ca65 is silly! It can export a `far' or `long' symbol if it is deduced
automatically, but specifying a `far' or `long' address size explicitely gives
an error! I mean, sure the 6502 does not have far or long addresses, but I only
want a symbol big enough for doing arithmetics or communicate with the linker
script, not using it as an actual address to load or jump to!