(WIP) Built using a makefile. New files.
Makefile is from boilermake, see on github.
It has some small changes to better support the cc65 suite.
main.mk is a file needed to Makefile to build the project.
test.sh is hence considered obsolete and is removed.

There's an init function that is called from reset, it is unfinished but it
initializes the BSS section and call constructors. Only the constructor
calls are tested for now.

foo.s tests some new features and foo2.s tests the pseudo-ops and constructors.

mapper69.ld is edited to support constructors, but the DATA section is also
moved. I'll see if there's a sane way to initialize the DATA section in RAM
while having it loaded in a bank other than the fixed bank. It is possible, but
it'll need further checks to verify that the segment didn't cross bank
boundaries.