Mapper control, safer nmi routine
New routines:  mapper_write  and  swprgbank_save.  The first only does a  mapper
write (X  ->  $8000,  A  ->  $A000), and  the second uses  the first  and  write
which PRG banks are mapped. The NMI routine restores the PRG banks at the end of
execution. Also a note is added in Calling_conventions.